+++ internal/pod/scaleconfig_test.go
@@ -17,12 +17,12 @@ limitations under the License.
 package pod
 
 import (
+	"errors"
 	"fmt"
 	"testing"
 
 	"github.com/ExpediaGroup/container-startup-autoscaler/internal/pod/podcommon"
 	"github.com/ExpediaGroup/container-startup-autoscaler/internal/pod/podtest"
-	"github.com/pkg/errors"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"k8s.io/api/core/v1"